SCA-based code improvements.

+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/oom/converter/ToDoubleConverterTest.java
@@ -22,11 +22,11 @@ class ToDoubleConverterTest {
 
     @Test
     void toAttributeSupportsWideningIntegerConversion() {
-        assertEquals(Double.valueOf(11), converter.convertToAttribute(Byte.valueOf((byte) 11)));
-        assertEquals(Double.valueOf(117), converter.convertToAttribute(Short.valueOf((short) 117)));
+        assertEquals(Double.valueOf(11), converter.convertToAttribute((byte) 11));
+        assertEquals(Double.valueOf(117), converter.convertToAttribute((short) 117));
         assertEquals(Double.valueOf(117), converter.convertToAttribute(117));
         assertEquals(Double.valueOf(117L), converter.convertToAttribute(117L));
-        assertEquals(Double.valueOf(3.14), converter.convertToAttribute(3.14F), 0.001);
+        assertEquals(3.14, converter.convertToAttribute(3.14F), 0.001);
     }
 
     @Test
