+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/oom/converter/ToShortConverterTest.java
@@ -24,11 +24,11 @@ class ToShortConverterTest {
 
     @Test
     void toAttributeSupportsWideningIntegerConversion() {
-        assertEquals(Short.valueOf((short) 11), converter.convertToAttribute(Byte.valueOf((byte) 11)));
+        assertEquals(Short.valueOf((short) 11), converter.convertToAttribute((byte) 11));
     }
 
     @Test
     void toAttributeSupportsIdentityConversion() {
-        assertEquals(Short.valueOf((short) 11), converter.convertToAttribute(Short.valueOf((short) 11)));
+        assertEquals(Short.valueOf((short) 11), converter.convertToAttribute((short) 11));
     }
 }
